buildAdTargeting test
